+++ example/lib/bloc/counter_bloc.g.dart
@@ -4,10 +4,7 @@
 // Generator: RxBlocGeneratorForAnnotation
 // **************************************************************************
 
-import 'package:example/bloc/counter_bloc.dart';
-import 'package:flutter/cupertino.dart';
-import 'package:rxdart/rxdart.dart';
-import 'package:rx_bloc/bloc/rx_bloc_base.dart';
+part of 'counter_bloc.dart';
 
 abstract class CounterBlocType extends RxBlocTypeBase {
   CounterBlocEvents get events;
@@ -20,20 +17,18 @@ abstract class $CounterBloc extends RxBlocBase
   ///region Events
 
   ///region increment
-  @protected
-  final $incrementEvent = PublishSubject<void>();
 
+  final _$incrementEvent = PublishSubject<void>();
   @override
-  void increment() => $incrementEvent.add(null);
+  void increment() => _$incrementEvent.add(null);
 
   ///endregion increment
 
   ///region decrement
-  @protected
-  final $decrementEvent = PublishSubject<void>();
 
+  final _$decrementEvent = PublishSubject<void>();
   @override
-  void decrement() => $decrementEvent.add(null);
+  void decrement() => _$decrementEvent.add(null);
 
   ///endregion decrement
 
@@ -45,10 +40,9 @@ abstract class $CounterBloc extends RxBlocBase
   Stream<String> _countState;
 
   @override
-  Stream<String> get count => _countState ??= mapToCountState();
+  Stream<String> get count => _countState ??= _mapToCountState();
 
-  @protected
-  Stream<String> mapToCountState();
+  Stream<String> _mapToCountState();
 
   ///endregion count
 
@@ -57,10 +51,9 @@ abstract class $CounterBloc extends RxBlocBase
 
   @override
   Stream<bool> get incrementEnabled =>
-      _incrementEnabledState ??= mapToIncrementEnabledState();
+      _incrementEnabledState ??= _mapToIncrementEnabledState();
 
-  @protected
-  Stream<bool> mapToIncrementEnabledState();
+  Stream<bool> _mapToIncrementEnabledState();
 
   ///endregion incrementEnabled
 
@@ -69,10 +62,9 @@ abstract class $CounterBloc extends RxBlocBase
 
   @override
   Stream<bool> get decrementEnabled =>
-      _decrementEnabledState ??= mapToDecrementEnabledState();
+      _decrementEnabledState ??= _mapToDecrementEnabledState();
 
-  @protected
-  Stream<bool> mapToDecrementEnabledState();
+  Stream<bool> _mapToDecrementEnabledState();
 
   ///endregion decrementEnabled
 
@@ -81,10 +73,9 @@ abstract class $CounterBloc extends RxBlocBase
 
   @override
   Stream<String> get infoMessage =>
-      _infoMessageState ??= mapToInfoMessageState();
+      _infoMessageState ??= _mapToInfoMessageState();
 
-  @protected
-  Stream<String> mapToInfoMessageState();
+  Stream<String> _mapToInfoMessageState();
 
   ///endregion infoMessage
 
@@ -99,10 +90,9 @@ abstract class $CounterBloc extends RxBlocBase
   CounterBlocStates get states => this;
 
   ///endregion Type
-  @override
   void dispose() {
-    $incrementEvent.close();
-    $decrementEvent.close();
+    _$incrementEvent.close();
+    _$decrementEvent.close();
     super.dispose();
   }
 }