+++ example/lib/bloc/details_bloc.dart
@@ -1,8 +1,10 @@
-import 'package:example/bloc/details_bloc.g.dart';
-import 'package:example/repository/details_repository.dart';
 import 'package:rx_bloc/annotation/rx_bloc_annotations.dart';
 import 'package:rx_bloc/rx_bloc.dart';
 import 'package:rxdart/rxdart.dart';
+
+import '../repository/details_repository.dart';
+
+part 'details_bloc.g.dart';
 
 abstract class DetailsBlocEvents {
   void fetch();
@@ -25,7 +27,7 @@ class DetailsBloc extends $DetailsBloc {
   DetailsBloc(this._detailsRepository);
 
   @override
-  Stream<String> mapToDetailsState() => $fetchEvent
+  Stream<String> _mapToDetailsState() => _$fetchEvent
       .startWith(null)
       .flatMap((_) => _detailsRepository.fetch().asResultStream())
       .registerRequest(this)