+++ CHANGELOG.md
@@ -1,3 +1,10 @@
-## [0.0.1] - TODO: Add release date.
+## [0.0.1] - Jan 16, 2020
 
-* TODO: Describe initial release.
+* TODO: Initial release.
+
+## [0.1.0] - March 15, 2020
+
+* Updated README file
+* Updated rx_bloc dependencies
+* Added UI Integration tests using Flutter Driver
+* Added Unit Tests using the rx_bloc_test package